Add files via upload
Virat-Kohli-ODI-Analysis
In this notebook. I have analysed the data of Virat Kohli's ODI Matches! the number of innings he has played, and the number of times he has been dismiised. The main motto is to analyse the grounds he has played in, and the type of runs he has scored in those grounds. With the analysis, we can predict the outcome of the next match. I Will keep working on the analysis and will be able to predict wheather the batsman would be able to score a century or not against a particular opposition using the classification algorithms. We can also predict the exact score using the regression techniques